clearview/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
motor==3.5.0
python-dotenv==1.0.1
pydantic-settings==2.5.0
pydantic[email]==2.9.0
google-generativeai==0.8.0
elevenlabs==1.9.0
stripe==10.0.0
python-multipart==0.0.9
pillow==10.4.0
httpx==0.27.0
